Move ConnectionTimeoutTestWithThread to manual test
